[REFACT]: refactoring
